temp/pyproject.toml
[project]
name = "fileagentwithmcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "deprecated>=1.2.18",
    "google-adk>=1.7.0",
    "lxml>=6.0.0",
    "mcp>=1.12.0",
    "requests>=2.32.4",
]
